+++ packages/solana_mobile_client/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0
+
+ - Bump "solana_mobile_client" to `0.1.0`.
+
+## 0.0.4
+
+ - **FEAT**: update mobile wallet client example (#667).
+
 ## 0.0.3
 
  - **FEAT**: update library to 1.0.0 (#647).